+++ jdk17u0117/src/hotspot/share/jtsan/lockState.hpp
@@ -44,7 +44,7 @@ class LockShadow : public CHeapObj<mtInternal>{
 
         Vectorclock* indexToLockVector(uint32_t index);
         
-        size_t getCurrentLockIndex(void);
+        uint32_t getCurrentLockIndex(void);
 
         void incrementLockIndex(void);
